[FIX] v2.OData(List|Tree)Binding: Correctly sort according to Edm type
When using clientside sorting, the Edm type was ignored and the default
comparator method of the SorterProcessor was used. This was wrong for
many types, especially  Edm.Decimal, where numbers are stored as
strings.

Moved default comparator function to Sorter class as public static
method and added the possibility to provide a custom comparator in the
Sorter constructor.

Closes https://github.com/SAP/openui5/pull/725
BCP: 1570881701
Change-Id: Iaebc3f8a11c61a80a346c4e9e0f464510dee7369