[INTERNAL] QUnit: make qunit-junit testDone cleanup more robust
If a global assert function is defined in an external script tag this
will finally cause the qunit-junit script to break in the testDone phase
when it tries to cleanup the window.assert object which has been set via
the external script. Only in this case the delete operator will run into
an exception whereas in any other case it returns a boolean value which
indicates whether the cleanup was successful or not.

The fix now also produces a failing test at the end which indicates to
the test creator that some script defines a global assert function which
needs to be checked. This failure could be overridden with a
QUnit.config.ignoreCleanupFailure.

Change-Id: I30aaa633ea26c6ef18ae85bdf43778eb30f71897
BCP: 1670101052
CR-Id: 012003146900000260212016