[INTERNAL] Module Loading: isolate times for lazy parsing of functions
Some Javascript engines support lazy function parsing / compilation.
When a function has not been parsed yet and is invoked for the first
time, the time needed for parsing / compilation is added to the self
time of the calling (!) code, at least in Chrome. This change makes this
more obvious for two prominent cases of lazy function parsing:

- when a preloaded module is executed
- when the factory of an AMD-like module is executed

An explanation of lazy parsing can be found at
http://ariya.ofilabs.com/2012/07/lazy-parsing-in-javascript-engines.html

Change-Id: Ic04365ebf2d59acf2e85700bccdb1c9c6caf35e4