[FEATURE] Model: new propertyChange event for property binding changes
The application can attach to this event on the model and gets notified
when changes occur to a property value in the model. The event provides
the parameter reason, path, context and value. Reason contains
information about the cause of the value change. Currently only reason
sap.ui.model.ChangeReason.Binding is supported which is fired when two
way changes lead to a value change of a property binding. The properties
path, context and value contain the property binding path, context and
its value.

Change-Id: Ibdee44fc5b7cdcc692dcc7875b31349161545062