[FIX] Popup: Order of Functions During Opening Fixed
There are 3 ways to open a Popup:
1. No animation
2. Default animation
3. Custom animation

For all cases is must be ensured that Popup.prototype._duringOpen is
really really called before Popup.prototype._opened.
Now the Popup ensures that the order is correct and thus private
functions '._opened', '_duringOpen', '_closed' and '_duringCLose' were
introduced.
Therefore these functions can be provided for the custom animation as
a callback with the corresponding scope of 'this' instead of annonymous
functions.
Plus, these private functions are testable instead of annonymous functions.

CR-Id: 012003146900000441282016
Change-Id: Iff613fe1e0f45db3c2dbe7a3944855c5f1ae76d6
BCP: 1670107952