[INTERNAL] ThemeVersion QUnit: enable parameters test
This test has been disabled with 62b1677c01b96e8284f81c6e0f10b75816dce5eb
and is now enabled again.
It was changed to override the "background-image" style of the
<link> tag to prevent usage of inline theming parameters as the
test only focuses on the library-parameters.json request.

Furthermore the check for an internal URI parameter to prevent the
usage of inline parameters has been changed to use
jQuery.sap.getUriParameters().

Change-Id: I0084137215146033e17412d47a1dc5b5df8527a3