[FEATURE] Component: added preload marker for models for manifest first
In case of manifest first scenarios models can be "preloaded" by
explicitly mark them. In the manifest the model should be flagged with
the option "preload": true at sap.ui5/models/XXX:

"sap.ui5": {
  "models": {
    "myModel": {
      "preload": true
    }
  }
}

Once the manifest is loaded by the Component factory it will immediately
trigger the creation of the models which are marked for preload. In case
of "preloading" a model you should ensure that the content of the model
(e.g. for ResourceModel, JSONModel or XMLModel) is not part of the
Component-preload as this would result in additional requests.

Change-Id: Ib66e9660779175e14d1131a0cd5ca211432b298f