[FIX] DataBinding: properly cleanup binding info when binding becomes invalid
When a binding is removed due to a binding context lost etc, the 
reference to the binding was removed from the binding info, but not 
the modelChangeHandler nor the dataChangestateHandler.

The same happens when an object with active bindings is cloned.

As a consequence, code that only checked for the existence of one of
these handlers failed (e.g. _handleLocalizationChange).

With this change, also the handlers are removed from the binding info. 

Change-Id: Ia9a18af40b8cfb66597958797aa8a0e1c20a823c