[FIX] ODataAnnotations: AnnotationParser - Error Messages in Debug Mode
The AnnotationParser showed error-messages if a Record Element 
contained PropertyValue as well as Annotation elements as they might
potentially overwrite each other in the resulting annotations-object.

The error message has now been changed to only show if an entry with
the same key already exists in the annotations-object and is thus in
fact overwritten.

The error is only displayed in debug mode; the condition will not be
checked in productive scenarios.

CR-Id: 012003146900000342682016
Change-Id: I97254244c5c1006063e8809d04eecb9b0319bfbe