[FIX] sap.m.OverflowToolbar - changing child visibility triggers invalidation
BCP: https://support.wdf.sap.corp/sap/support/message/1670176867
This change fixes a regression introduced with an optimization to the method that calculates the width of child controls. The optimization allowed the method to return the last known width of controls that are now in the overflow (thus not in the DOM) by providing a second parameter and passing the last-known width with it.
However, an invisible control was incorrectly reported with its old width, and not 0, which caused wrong overflow calculations in the end.

Change-Id: Ide926d57c0d092aebe608db4fbb1a71f9c9be6ab