[FIX] ManagedObject: Propagate properties when setting parent to null
When a control was removed from the control tree, the bindings for
inherited models were not properly destroyed and still active.
To fix this, bindings are now updated and model information
propagated,then the parent is set to null.

In case of a 'move' - remove/add controls snychronous in an 
aggregation - we should not propagate synchronous when setting the
parent to null. Synchronous propagation destroys the bindings
when removing a control from the aggregation and recreates them when
adding the control again.
This could lead to a data refetch, and in some scenarios to endless
request loops.

Change-Id: If848da4e19910c86b617fee4ebac33c2a08aa364
CR-Id: 012003146900000317262016
BCP: 1670131259