[FIX] Component: destroy preloaded models in case of errors
In case there is some error while preloading / creating a component
instance via "sap.ui.component" it should be made sure that all
premilinary created models will be destroyed.

Refactored internal "loadComponent" method to allow passing options
depending on the usage (sap.ui.component/sap.ui.component.load).
This prevents the issue with some "asyncHints" which are only
supported in one of the methods (e.g. waitFor and preloadOnly).

Moved handling of "asyncHints.waitFor" promise to "loadComponent"
to be able to react on a rejection to destroy the models.

Enabled "async" preload mode for QUnit tests to be able to test if
models will be destroyed in case a library can't be pre-loaded.

Change-Id: Ibafc9ad387bab80b9f5223c31ca5806ff1f77fdd