[FIX] ImageMap: Use Selector As Written In The Spec
Due to the upgrade of jQuery to 2.2.2 there were issues with the selector
for the elements inside the map.
We checked the W3C-spec and found out that the selector 'usemap' has to be
used without capital letters, that the quotes are necessary and that the '#'
has to be used as well.
So we made this consisten throughout the entire framework.

Change-Id: I7955dddfadc642caefd0eecc402f2b179e2ce6b2