[FEATURE] Service Injection: central registry and Component usage
The Service Injection concept allows to register so called Service
Factories in the Service Factory Registry. A UI5 Component can declare
the usage of a Service within the manifest: 

{
	...
	"sap.ui5" : {
		"services": {
			"MyServiceAlias": {
				"factoryName": "my.ServiceFactoryAlias"
			}
		}
	}
	...
}

Once the Service is requested in the Component via a getter API the
Component internally looks-up the Service Factory in the Service Factory
Registry and creates a scoped Service instance with the context of the
creator (Component). This is useful to avoid passing this context later
on when using the Service instance in the Component itself.

Change-Id: I422bc9de4b5a0d3288fb762c6c926939826e9b4d