[FIX] sap.ui.core.util.File: handle utf-8 content on Safari
This commit fixes issues when saving files with utf-8 content
on Safari. "btoa" can not handle utf-8 characters and therefore
threw an exception.
Encoding the content beforehand would result into an encoded file
and the user would have to manually decode the downloaded file
which would not be an option.
Therefore the content will only be urlencoded but not base64 to
solve issues with utf-8.

BCP: 1570888105
Change-Id: I84c2c89efede9a2d01f79a618c3b8f4e27b48f0b