[FIX] Service Injection: optional flag, anonymous services, JSDoc, fixes
Optional flag in Descriptor for Components (manifest.json):
New property: "sap.ui5/services/*/optional". getService returns a
Promise which provides the service interface when resolved. The optional
property defines that the service is not mandatory and the usage will
not depend on the availability of this service. When requesting an
optional service the <code>getService</code> function will reject but
there will be no error logged in the console.

Anonymous services (shorthand):
To simply define and register Services and ServiceFactories a shorthand
function has been added:

var oServiceFactory = new ServiceFactory({
  init: function() {},
  exit: function() {},
  doSomething: function() {}
});

Change-Id: I777e2f10af36178972535fd456915dbeb236355e