[INTERNAL] Debug mode: allow to load debug sources selectively
The URL parameter 'sap-ui-debug' now supports a comma separated list
of 'glob' patterns. For modules whose name matches one of the patterns, 
their debug source will be loaded. The previously supported values
'true', 'x' and 'X' work as before: they completely suppress preloads
and load debug sources for all modules.

Example:

  index.html?sap-ui-debug=sap/ui/model/odata/v2/
 
will load all debug sources for all modules of the OData V2 model.
All others modules will be taken from the preload (if preload is
active).

As a POC, the TechnicalInfo popup now allows to open a dialog to
interactively select packages / modules for which debug sources should
be loaded. The result will be persisted in local storage as before.
If the feature proves to be helpful, the support tool and/or the Chrome
extension should be enhanced.

Change-Id: Ibc804d2e805f38ca72dcf62b62e8e05bcad85d2f