[FEATURE] model.odata.AnnotationParser: Support Nested Annotations
Up until now, the AnnotationParser did not support parsing of nested
annotations (meaning Annotation nodes inside other Annotation nodes).
Now we search for further annotation nodes that are inside the top
level annotation. In the resulting annotations object, these 
annotations are then also located under the parent annotation.

In order to support this kind of nesting the parser had to be 
refactored in order to allow access to the used annotation key from
inside the value parsing step.


CR-Id: 012003146900000436102016
Change-Id: I3adb3c91193e9cb3c4861823517458d5fbec1e5f