[FEATURE] sap.m.Select: provide fast item navigation feature
Provide fast item navigation feature to improve usability on
desktop environments. The fast item navigation feature is a
convenient way of selecting an items in the dropdown list.
For example, in a list of countries, if you want to select
a specific item, you can fast type the starting character of
the item's text that you want to select and the next item
matching the typed character would be selected automatically.

Change-Id: I026684be8c8323c7573be4627b76dac2a6dfd4a2