[FEATURE] DateFormat: Allow mixed styles for datetime format
According the CLDR it is allowed to have mixed datetime formats, e.g.
long date and medium time. The DateFormat API did not allow that, the
style given for datetime formats was used for both date and time style.

This change introduces the possibility to defined mixed styles
("medium/short" or "long/medium") for datetime formats.

BCP: 1670235968
Change-Id: I4fb47d954c697c4449f028c6a960de202dcf88c8