[INTERNAL] Component: rebuild manifest loaded callback to load component callback
initial Change-Id: I276955558d8ba21b22f886f4b78ec48e4c62f5eb
The preliminary callback handler provides access to the configuration
after the load component was triggered.
It replaces the preliminary callback called after the manifest was
loaded due to the fact that the consumer can already work with the data
present in the configuration.
In addition the passed startup parameters within the configuration
prohibit a unnecessary backend call if the "workaround of app variants
with the startup parameter sap-app-id" is present.
This hook will only
be used by UI flexibility and will be replaced with a more generic
solution in future.

Change-Id: Ia9eefd001b689f9e7d11860ed807861bd3e0b2c9