[INTERNAL] sap.ui.commons.Tree - Filtering of the tree fails when bindAggregation with a template is used to create the tree
- In updateNodes method the wrong getter was used to get the nodes returning the new nodes rather then the current ones
and on a corner case with filtering of the model a new node was destroyed instead of old one resulting in javascript error and
misleading result. This behavior is observed only when the nodes are bound using bindAggregation method.

BCP: 1670267206

Change-Id: I4b34a0c009bac7c35f6a0d1babc0370bbb581ab4
(cherry picked from commit 48e133b64a553a47145817fac00f96476a0791e4)
CR-Id: 012003146900000460652016