[FIX] ManagedObject: Propagate properties when setting parent to null
In case of a 'move' - remove/add controls snychronous in an 
aggregation - we should not propagate synchronous when setting the
parent to null. Synchronous propagation destroys the bindings
when removing a control from the aggregation and recreates them when
adding the control again.
This could lead to a data refetch, and in some scenarios to endless
request loops.

Change-Id: Ib2ddb387a06ac63cc56613669ab35199a9aefe6e
CR-Id: 012003146900000464992016
BCP: 1670131259