[INTERNAL] LessSupport: no last-modified for less means prefer less
If the last modified check of the less file returns 0 (not present) the
less files should be used instead of the CSS files. E.g. in Web IDE
scenario the less files from the local workspace will only have an ETag
but no last modified header and this fix now corrects the behavior that
the less files will be used instead.

Change-Id: I4cd18f3480795426a7c311423834af2ce7f782a6
CR-Id: 012003146900000475592016