[FIX] sap.ui.table.Table: register resizeHandler synchronously if table has no size after rendering
- when the table is used in a popup, it may happen that the tables onAfterRendering is processed while
the popup has still display:none. This results in a table size of 0x0. In the next browser painting though, the
popup is set to display:block but since the tables ResizeHandler is registered in a delayed call, it's just too late
to encounter the visibility change of the popup.

Change-Id: I36f56c7301602de5f889df815bf2eb0c126b0047
BCP: 1670264580
CR-Id: 012003146900000478562016