[INTERNAL] ListBinding: Followup fixes for the new diff calculation
The function to determine the key should get the binding context,
instead of the data object. This allows a relative binding path to be
provided for the key, instead of just a property name.

Use a try/catch for the diff calculation in JSONListBinding to avoid
issues with cyclic references in the JS object treee.

Change-Id: I6b74d0f838d1078a7a4a196341154ec4dae3f608