[FIX] sap.m.DatePicker: invalid input lost after rerendering
If a user enters some invalid text in the DatePicker the field is marked
as invalid. After re-rendering the last valid date is rendered, but the
DatePicker is still marked as invalid.
Now the invalid value is rendered again to keep the field stable.

If the value property is bound to a model and while updating the value
the models formatter cahnges the value the wrong value is used inside
the cahnge event.

Change-Id: I823b48bc6debd6ad708c01cca740227f3cd44c61
BCP: 1670279672
CR-Id: 012003146900000484582016