[FIX] sap.m.FacetFilter: remove facet icon is closed when mouse is really released over icon.
When clicked on the remove facet icon, the opened popover is closed (due to focusout). As result in it's close handler
the icon above is removed/hidden.

This has a side effect of when the user clicks and holds the mouse button for a few seconds - the popover closes
and the icon is removed even though the user has'not released the mouse button yet.

On the other hand the flow 'popover close->X button remove' is valid when the user clicks outside the popover, but not on the remove facet icon.

The fix decouples the flow to touchstart, touchend, press, then differentiates both scenarios and prevents such icon removal.

(cherry picked from Id85d667532b62faa381ee6bbb452fb44b934b30c)

Change-Id: Icdba17e7c075305310ba38affa6c206e33f2f4a1
BCP: 1670213717
CR-Id: 012003146900000491562016