[FIX] core.Popup: Adding Public Methods in PopupSupport Fixed
Due to performance improvements it isn't checked anymore if a method
that should be added to the control's metadata already exists.
Since the metadata is shared between all control instances it is possible
that a public method is added multiple times. Only by calling 'getPublicMethods'
on the control duplicates within the array will be removed.

Thus the PopupSupport now checks if a function that should be added
to the metadata already exists and skips adding the methods.

Change-Id: I80ba470e18a62805b5b16c84624e193704dae06a
BCP: 1670265672