[INTERNAL] Core: harmonize metadata for special settings
Metadata for special settings was not consistent between runtime and
documentation (JSDoc). Default property now consistently is 'type' (was
'readonly' before in JSDoc, but 'type' in runtime).

Special settings furthermore support a visibility ('hidden' or 'public',
default is 'public').

Existing special settings have been adapted to the new conventions.

This change needs to be seen in conjunction with jsdoc-maven-plugin
version 1.1.18 or higher.

Change-Id: Ic75972f99e046c4d7644cc3d89569416479ffda6