[FEATURE] New Explored examples for TreeTable with OData
"Hacking the Explored App" TreeTable examples from UI5con 2016

With an OData model, several requests for each node must be sent in
order to expand them. If batch-requests are enabled it will be one
batch for each auto-expand-level but the batch requests will contain
sub-requests for every single hierarchy node.