[FEATURE] New Explored examples for TreeTable with OData
"Hacking the Explored App" TreeTable examples from UI5con 2016

With an OData model, several requests for each node must be sent in
order to expand them. If batch-requests are enabled it will be one
batch for each auto-expand-level but the batch requests will contain
sub-requests for every single hierarchy node.

Change-Id: I7a2cf506d996491209732010402aa6ac090b9eb6
Closes https://github.com/SAP/openui5/pull/994