[FIX] Serializer: fix handling of property default values
If the metadata for a managed property specifies defaultValue:null, 
then this means: property has no own default value, fall back to the
default value of the type of the property (e.g. 0 for type int). 
It does not mean: use 'null' as default value.

Both, the HTML and XML serializer ignored this specific meaning of
'null' and therefore sometimes unnecessarily wrote out the default
value. Instead, they ignored any null or empty value, even if that was
not a default value.

Furthermore, default values only exist for properties. There should be
no such check for associations and aggregations.

Change-Id: I22490d81ea52c8774f42619f5b0fadfe25f84054