[INTERNAL][FIX] TemplateControl: make cleanup more robust
Due to the recent changes in model context propagation, the model
might already be removed from the control when destroy() is called.
Cleanup therefore shouldn't try to receive the model information from
the control, but use the information stored in the binding.

Also handle the case that no binding has been created (e.g. when no
model was available during .bind)

Change-Id: I41c78431f7a4053e448c83b16ba6490dc21788cf
BCP: 1670276051