[FIX] core.BusyIndicator: Removed Vendor-Prefixes for Animations
As from Internet Explorer 10 and newer Firefox, the vendor prefix
for keyframes animation isn't needed anymore it could be removed.

On the other hand using attributes within an animation with unknown
vendor prefix in Internet Explorer, the browser crashed and was also
not able to fetch the stylesheet at all.
So a removal was necessary for this specific case.

Even though Internet Explorer 9 isn't able to handle keyframes
animation the workaround for the BusyIndicator remained.

CR-Id: 012003146900000534562016
Change-Id: I91d375c5d0b8976f3f4f5386acb3a4c2a8630380
BCP: 0020283765 0000224717 2016
BCP: 0020283765 0000246537 2016
BCP: 0020283765 0000192865 2016