[FIX] v2.ODataModel: Allow custom Content-Type header for media entities
To prevent the Content-Type header to be set by datajs automatically, we
did always set the Content-Type header ourselves in the ODataModel. This
prevented custom Content-Type headers from the application to be
accepted, which are needed to create POST/PUT requests for media
entities, which can have arbitrary content types.

BCP: 1670308359
Change-Id: If93fcb2dc759a7dbefa30bec28a42f0f300103a1