[FIX] sap.m.Table ACC demo: fix sap.m.Select tooltip reading issue
By default, the sap.m.Select text field automatically size itself
to fit its content. To archive this auto sizing feature,
it is required to render a shadow DOM (invisible element)
inside the sap.m.Select text field.
As a side effect JAWS16 + IE11 don't announce the tooltip and
instead announce the text field content.
To work around this issue, the width of the sap.m.Select is set
to 11rem, this avoid the required shadow DOM to be render.

Change-Id: I8057d09d3c22b285448bc7be3bee37cb9752547d
BCP: 1670314301