[FIX] Core: add commonly used 3rdparty to preload, remove jQuery-2.1.4
Since routing meanwhile is used quite frequently, the necessary
thirdparty components 'signals', 'crossroads' and 'hasher' should become
part of the sap.ui.core library-preload bundle. 

As the same configuration is used for preload and the sap-ui-core-all.js
file, the mentioned components are also included in the 'all' file. 
To make this work, module loading has to the determine the exports of
such 'raw' modules even when they are loaded without jQuery.sap.require. 

Last but not least, jQuery-2.1.4 has been removed from the preload
bundle, it was added only by mistake (no need to preload jQuery, jQuery
is a precondition for UI5 and must have been loaded already)

Change-Id: Ie18cc1232d06f03c26d8982f2d51f82de30245b7
CR-Id: 012003146900000559542016