[INTERNAL][FIX] Robust UI5 TestRunner.js test completion check
The UI5 Test runner was checking for test completion based on the test title
text. This was causing false positives, in cases where the word "completed"
appeared in the title of a test. This fix checks for the qunit-banner status
(pass or fail), which should be a more reliable indicator of test results.

PS1: main change
PS2: replace spaces with tabs

Change-Id: Ie0ea7f28147da483bd757f7bff2673bb10b4a878