[FIX] sap.ui.layout.ResponsiveFlowLayout: when used in SimpleForm items' height is calculated consistent for all browsers.
In IE height/width is reported as floating point number. Some checks 
heavily rely on comparision of those values in order to decide for 
further (re)rendering.
This fix makes sure all sizes have the same scale(precision) so the 
comparison is not affected by any fluctoation.

For example: IE reports element height 51.012349<8>714, previous was 
51.012349<9>714. This should not be an issue/reason for rerendering, 
but actually was. 
After rerendering the next measarement slightly differs again, which in 
turn was causing rerendering. This turns out in an endless loop of rerendering.

Change-Id: I0e6a304a185d804673ddbc7ed141c6303ea5c14a
BCP: 002075129500001241092016