[INTERNAL] v4._ODataHelper.createCacheProxy: Refactor
Prepare to introduction of filter: createCacheProxy now takes a promise
which resolves with the canonical path to create the cache.

In a later change we will enhance createCacheProxy to take an additional
promise which resolves with the $filter value for given filters so that
this can also be used for cache creation.

PS1: PP with KTH
PS2: comments SBE (createCacheProxy signature simplified)
PS3: minor comments
PS4: rebased
PS5: review comment

Change-Id: If0456a6a4d4a8df1ac5eaffa77d24a31df5b432f
JIRA: CPOUI5UISERVICESV3-60