[WIP INTERNAL] sap.ui.core.theming.Parameters.getActiveScopesFor
Some controls do render the scope classes through the render classes
instead of using the Custom Style Class API functions, what implicates
that scope classes could possibly be set on child/inner DOM-Elements which could not
be recognized by the actual concept.

Due to this we need to adopt "Parameters.getActiveScopesFor" and enhance the lookup with navigating the DOM hierarchy, if the DOM is available. Otherwise we navigate the control chain.

Change-Id: Iaa34ac0bc6e2c045ca769e85077d0d7ffddecc96