[INTERNAL] Component._fnLoadComponentCallback called after libs-loading
The sap.ui.core.Component._fnLoadComponentCallback was not within the
promise-chain until now.
This caused a race condition: If the callback is registered within a
currently loading library the check for the callback sometimes was to
early and the registration was not yet done.
This change now postpones the callback until the async hint library
loading is finished and therefore the race condition is gone.

PS 5: reverted PS 4

Change-Id: I54cefb24e18c6c689fac7d5dee6aa8da92e443aa