[FEATURE] jquery.sap.interaction: identify loading component
As Tiles in FioriLaunchpad trigger the loading of a new component, up to
now the component in the interaction measurement was the FioriLaunchpad.
This is actually not correct, as the efforts following the Tile click
belong to the newly loaded Component. Hence an API needs to be called,
which communicates the name of the newly loaded component to the
interaction measurement, so that the correct one can be displayed.

CR-Id: 012006153200001377792016
Change-Id: Ic3064952efb28177d1325d4decd1bec97c0c9c0f