[INTERNAL] ODataListBinding: parameter for valid dataReceived events
The dataReceived event does not send any parameters in the error/abort
scenarios, which is being used to ignore such events from valid ones.
However, in some cases, there is no parameter sent with the event which
causes some UI/apps to ignore such seemingly "valid" events and hence
behave incorrectly.

The behaviour is worse for the old (non-v2) ODataListBinding where a
filter request is not event being sent (when count is 0 - pretend mode)
and the UI shows invalid information. Until a better fix is done (or all
apps move to v2.ODataModel), this can be used to resolve the issue.

BCP: 1680054608
CR-Id: 012006153200001209972016
Change-Id: Ifae31f39c758aa5343551a5b9ba2cf1ad2f220ee