[INTERNAL][FIX] sap.m.Fiori20Adapter: Caching no longer stops secondary adaptation
The adapter caches adaptation results in order to prevent second (redundant) adaptation upon re-visit of an already adapter view.
This however, also stops adaptation when the component is re-lanched (and the control tree constructed from scratch). Therefore restricted caching to be used only within the current call to the adaptation function.

Change-Id: Ia9021574ab69ac86b6850ce21b80a5641f39fece