[FIX] Core: fix preload.json based fallback for library loading
The existing library-preload.json files store dependencies in the format
'my.other.lib.library-preload', but the new implementation as well as
the new testdata assumed the format 'my.other.lib'.

As the new format is mandatory to be able to trigger a
library-preload.js(!) preload first, dependencies from a
library-preload.json first have to be normalized to the new syntax.

Testdata (and the generation script) also has been fixed.

Change-Id: Id357859ee74cab30657550f0616a6e358d36da30
BCP: 1680069570