[INTERNAL][FIX] LibraryInfo: support patterns for ACH assignment
For some libraries (esp. sap.ui.core), the mapping from API name to ACH
(application component hierarchy, software component as used when
creating incidents via support.sap.com / BCP) is not 1:1. Furthermore,
sap.ui.core uses patterns to describe the mapping, not only full names.

The code in LibraryInfo ignored pattern entries in the .library files
and did not implement patterns in the mapping. Both issues have been
fixed: patterns in .library have been migrated to modules, LibraryInfo
also implements pattern matching, but only a trailing '*' (which should
be sufficient).

As a result, component calculation works again in the Explored app as
well as in the team app.

The descriptor for libraries (manifest.json) and its generation out of
the .library file is not affected as the descriptor only knows the
default component for a library, no exceptions for individual modules. 

Change-Id: I386e08fe3791bf500226e5aa7c606f434799001e