[FIX] Icon Tooltip: Prevent using technical names
Technical icon names have been used as tooltip fallback in case
no tooltip was provided or the icon is not translated.
As it confuses users this fallback has been removed.

The "aria-label" attribute still falls back to the technical name
to prevent a browser fallback to e.g. Emoji icon names which would
be even more confusing.

Adopted Icon QUnit test to use an icon with translation.

CR-Id: 012006153200001229172016
Change-Id: I874de520406d68e5d724bdaa3ab3b4cb5466c7b5
BCP: 1680063559