[FIX] sap.ui.core.Popup: blind layer iframe issue
The Popup uses an internal "blind layer" with an <iframe> to work
around an IE issue that existed from IE 8-10 showing embedded
content (like flash) above the popup content which is not expected.

The "src" attribute causes issues e.g. with Windows 8.1 Apps
and can be safely removed as the default content is a blank html
document.
As the blind layer isn't even required for IE 11 (which is used in
Windows 8.1 Apps) the issue is already solved by adopting the
if-clause to also check the browser version.

But to make sure the iframe doesn't cause any other issues both
changes have been done.

CR-Id: 012006153200001242382016
BCP: 0020751295 0000133662 2016
Change-Id: I03ebee4e5f33223f6ad5cd5505870628d91334b3