[INTERNAL] Theme Build: CSS Scope Root rule
Regarding the change (7eb6df2fe41b6cb101785d44d8530798c7fad154) specific
prescoped css rules for the 'sapUiBody' class were added to the
shared.less.

In order to have a generic solution, this prescoped rule regarding
'sapUiBody' is replaced by a custom css rule '#CSS_SCOPE_ROOT'.

This custom rule will be handled by the merge in the sap.ui5.runtime
theme build and replaced with the respective scope name, as the
specifity of this new rule stays the same. Furthermore, this custom rule
allows it to propagate property values not only for sapUiBody.

Change-Id: I30202c6dcb5b4de006df2c489f896f0f9f48c2cb