[INTERNAL][FIX] Demokit: fix handling of pages that are not indexed
When a page (hash) was not listed in any index, no index (left hand 
side tree) was displayed. But visually, the 'Developer Guide' seemed 
to be selected. Even worse, it was not possible for the user to select 
the 'Developer Guide' with the next click, only after selecting some
other index.

With this change, the 'Developer Guide' is displayed as a fallback when
no other index matches the current page. This applies to unindexed 
pages as well as to missing pages (404s).

Change-Id: Ic820feb8a69229145636e06f5f01abcb7990f24d
CR-Id: 012006153200001249272016
BCP: 1570783625