[FIX] XMLTemplateProcessor: Ignore empty binding property
In the smart templating it is not possible to have optional attributes
on tags, but the binding for the attribute value can just return an
empty string. If the templating generates binding="" it should be
treated as if no binding was defined.

BCP: 1680072983
Change-Id: I57cacd0667e73b7cd7c08b9ac038e2ecb2f852e6
CR-Id: 012006153200001280652016
(cherry picked from commit 47e3d4b411c045f62c53da0f44a5b27b78d61423)