[FIX] core.BusyIndicator: Fixed positioning of animation for RTL-mode
The animation for the busy indicator was not positioned properly
for right-to-left (RTL) mode. Before the change the animation was
positioned at the left border of the screen.

Now the animation is centered for right-to-left and left-to-right.

BCP: 1670327378
CR-Id: 012006153200001301922016
Change-Id: Ide2e823d9df17b9684cf95e90b8e594c2e9934d4
(cherry picked from commit 61baaeaaa0a8616b18bfe5e034fcd5fb611a1059)