[FIX] core.Popup: Check in 'onAfterRendering' extended
When a DOM element was set as content to the Popup and the content
was rerendered it crashed in 'onAfterRendering' because it was assumed
that only control can be rerendered.
If a control doesn't set its instance as content but its DOM reference,
it caused the Popup to crash when the contol was rerendered.

Now in 'onAfterRendering' it is checked whether a control or DOM reference
is used as content.

CR-Id: 012006153200001304742016
Change-Id: Ib416d11440ff3e13ff2d907ce70e6f422f81f359
BCP: 1670252630
(cherry picked from commit c90d49c96a9cb853499b81a059518b8cfe7e7793)