[FIX] sap.ui.unified.Menu: current item of type MenuTextFieldItem does not lost focus in IE and Edge anymore.
On selecting MenuTextFieldItem and moving out of it without releasing the mouse,
the item was loosing it's focus so the selected text was not editable anymore under IE and Edge.
On the rest of the browsers, the focus is still inside the item and any keyboard events are delivered on the item.

Change-Id: I95fc35948bbff736e80a24b2d73302da26d5804b
BCP: 002075129400001795982016
CR-Id: 012006153200001298652016