[INTERNAL][FIX] Core: accept conflicting sync and async preload requests
When a library should be preloaded synchronously while an asynchronous
preload request (<script> tag) for the same library is already pending, 
just accept and execute the sync requests. This leads to double loading
for the same JS file, but prevents errors in launchers that load
components or libraries twice, async and then sync.

Change-Id: I57e0842cfb75008fb8bc3d90d8b3497dddcb6350
BCP: 1680124328