[INTERNAL][FIX] No switch to sticky mode if no content is scrolled
When an instance of ObjectPageLayout is given display: "none", size recalculations are activated that change the page state in a way inconsistent with the initial state. In particular, the mode is changed from expanded to sticky. Fixed by making the check for switch to sticky mode more presize. No switch to sticky mode during scroll if all page content is still expanded (i.e. if scrollTop stil == 0).

Change-Id: I2e578840ba8ecaf7ea945a8cba91d871ae64c9cd
BCP: 1680061873