[FIX] v2.ODataModel: Call correct request handlers
There was an issue with merged same requests (single and batch), which were
aborted after send. When the first request is aborted, the errrorhandler
of all merged request was called. Also the aborted flag was reseted to
often. This caused issues in the ODataListBinding, where current requests
are aborted when refreshing and filtering.

Change-Id: Iabe9a468263f6cab8db95d301976bde062bde783
BCP: 1680110982