[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Section title complex binding no longer fails
Complex binding was not working in all cases, because the cloned binding info from the section title into the corresponding anchorBar button was incomplete. However, cloning of binding info is not even needed, because currently the entire anchorBar is rebuilt every time a section title is updated (i.e. ObjectPageSectionBase.prototype.setTitle is called). Removed the cloning and covered with unit tests for data binding of section title.

Change-Id: I1510e9fb4c50f37d4716197088e7b19a6225b779
BCP: 1680163929