[FIX] jQuery.sap.dom: first/lastFocusableDomRef must check CSS visibility
The old implementation of jQuery#firstFocusableDomRef and 
jQuery#lastFocusableDomRef did not take the CSS property 'visibility'
into account for the returned elements. So the functions might have
returned non-visible elements as 'focusable', which was wrong.

In addition to checking that property, the code has been improved to
achieve better code reuse and to avoid unnecessary jQuery instances. 
 
Change-Id: I24e70530bd01d06c131b57e70780622c487ad8b4
BCP: 1680044400