[INTERNAL] sap.uxap.ObjectPageSubSection - demokit sample updated
This change concerns the demokit sample that showcases the subsection's toolbar.
The sample did not present the case of a single subsection with a toolbar,
which is promoted to a section and the toolbar is visible on section level.
The aim of this change is to demonstrate exactly this behavior.

Change-Id: I1ccb33419c03bc1fdd3916024947f6792e29e77e